# specs/001-labor-daily-report/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Labor Management System API
  description: |
    API for managing daily contractors, work reports, and wage calculations.
    Supports 8 user roles with Thai language data.
  version: 1.0.0
  contact:
    name: Labor Management Team
servers:
  - url: https://api.labor-system.example.com/v1
    description: Production
  - url: https://staging-api.labor-system.example.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
...
security:
  - BearerAuth: []
paths:
...
  /wage-periods:
    get:
      tags: [Wage Periods]
      summary: List wage periods
      responses:
        '200':
          description: List of wage periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WagePeriod'

    post:
      tags: [Wage Periods]
      summary: Create wage period
      security:
        - BearerAuth: []
      x-required-roles: [Admin, PM, PD, MD]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WagePeriodInput'
      responses:
        '201':
          description: Wage period created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagePeriod'
  /wage-periods/{id}/calculate:
    post:
      tags: [Wage Periods]
      summary: Calculate wages for period
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Calculation complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WagePeriod'
...
  /scan-data/import:
    post:
      tags: [Scan Data]
      summary: Import scan data from Excel
      security:
        - BearerAuth: []
      x-required-roles: [Admin]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Import successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
  /scan-data/discrepancies:
    get:
      tags: [Scan Data]
      summary: Get scan data discrepancies
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [Type1, Type2, Type3]
        - name: status
          in: query
          schema:
            type: string
            enum: [Pending, Verified, Fixed]
      responses:
        '200':
          description: List of discrepancies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScanDataDiscrepancy'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WagePeriod:
      type: object
      properties:
        id:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        projectId:
          type: string
        status:
          type: string
          enum: [Draft, Calculating, Completed]
        dcSummaries:
          type: array
          items:
            $ref: '#/components/schemas/DCSummary'
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        calculatedAt:
          type: string
          format: date-time
    DCSummary:
      type: object
      properties:
        dcId:
          type: string
        employeeId:
          type: string
        name:
          type: string
        regularHours:
          type: number
        otHours:
          type: number
        regularPay:
          type: number
        otPay:
          type: number
        professionalFees:
          type: number
        totalIncome:
          type: number
        totalExpense:
          type: number
        socialSecurity:
          type: number
        netPay:
          type: number
    WagePeriodInput:
      type: object
      required: [startDate, endDate, projectId]
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        projectId:
          type: string
    ScanDataDiscrepancy:
      type: object
      properties:
        id:
          type: string
        employeeNumber:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
          enum: [Type1, Type2, Type3]
          description: |
            Type1: Daily Report < ScanData
            Type2: Daily Report exists but ScanData missing
            Type3: ScanData exists but Daily Report missing
        dailyReportHours:
          type: number
        scanDataHours:
          type: number
        status:
          type: string
          enum: [Pending, Verified, Fixed]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...